nvbios/bit: Correct size of BIT Header version field to 16-bits
Version field of BIT Header is actually 16 bits, with the major version
in the upper byte, minor version in the lower byte.

Until now, envytools was only reading the major version number.

Source: https://download.nvidia.com/open-gpu-doc/BIOS-Information-Table/1/BIOS-Information-Table.html

+++ nvbios/bios.h
@@ -81,7 +81,7 @@ struct envy_bios_bit_entry {
 struct envy_bios_bit {
 	uint16_t offset;
 	uint8_t valid;
-	uint8_t version;
+	uint16_t version;
 	uint8_t hlen;
 	uint8_t entriesnum;
 	uint8_t rlen;
